+++ tests/testthat/test-abbrev.R
@@ -1,5 +1,7 @@
 tm <- league_teams("42654852")
 
+# team --------------------------------------------------------------------
+
 test_that("team abbreviation", {
   x <- team_abbrev(1, tm)
   expect_length(x, 1)
@@ -16,9 +18,10 @@ test_that("team un-abbreviation", {
 
 test_that("team abbreviation errors", {
   expect_error(team_abbrev(1, mtcars))
+  expect_error(team_unabbrev("test", tm))
 })
 
-# -------------------------------------------------------------------------
+# slot --------------------------------------------------------------------
 
 test_that("slot abbreviation", {
   x <- slot_abbrev(0)
@@ -36,9 +39,10 @@ test_that("slot un-abbreviation", {
 
 test_that("slot abbreviation errors", {
   expect_error(slot_abbrev("test"))
+  expect_error(slot_unabbrev("test"))
 })
 
-# -------------------------------------------------------------------------
+# pos ---------------------------------------------------------------------
 
 test_that("position abbreviation", {
   x <- pos_abbrev(1)
@@ -54,33 +58,46 @@ test_that("position un-abbreviation", {
   expect_equal(x, 1)
 })
 
-# -------------------------------------------------------------------------
+test_that("position abbreviation errors", {
+  expect_error(pos_abbrev("test"))
+  expect_error(pos_unabbrev("test"))
+})
 
-test_that("position abbreviation", {
+# pro-team ----------------------------------------------------------------
+
+test_that("pro-team abbreviation", {
   x <- pro_abbrev(1)
   expect_length(x, 1)
   expect_s3_class(x, "factor")
   expect_equal(as.character(x), "Atl")
 })
 
-test_that("position un-abbreviation", {
+test_that("pro-team un-abbreviation", {
   x <- pro_unabbrev("Atl")
   expect_length(x, 1)
   expect_type(x, "integer")
   expect_equal(x, 1)
 })
 
-# -------------------------------------------------------------------------
+test_that("pro-team abbreviation errors", {
+  expect_error(pro_unabbrev("test"))
+})
 
-test_that("position abbreviation", {
+# stat --------------------------------------------------------------------
+
+test_that("stat abbreviation", {
   x <- stat_abbrev(1)
   expect_length(x, 1)
   expect_equal(x, "PC")
 })
 
-test_that("position un-abbreviation", {
+test_that("stat un-abbreviation", {
   x <- stat_unabbrev("PC")
   expect_length(x, 1)
   expect_type(x, "integer")
   expect_equal(x, 1)
+})
+
+test_that("state abbreviation errors", {
+  expect_error(stat_unabbrev("test"))
 })