+++ NEWS.md
@@ -1,3 +1,5 @@
+# fflr (development version)
+
 # fflr 2.1.0
 
 * **Package has been updated for the 2022 season!**