+++ NEWS.md
@@ -1,5 +1,8 @@
-# fflr (development version)
+# fflr 2.2.0
 
+* **Package has been updated for the 2023 season!**
+  * Functions now use `seasonId = 2023` by default in `ffl_api()`.
+* Update objects with latest data from 2023 season.
 * Fix `nfl_players` to include actually all of the players.
 * Adjust the free agent information in `nfl_teams`.
 * Add more function tests.
@@ -7,8 +10,7 @@
 
 # fflr 2.1.0
 
-* **Package has been updated for the 2022 season!**
-  * Functions now use `seasonId = 2022` by default in `ffl_api()`.
+* Functions now use `seasonId = 2022` by default in `ffl_api()`.
 
 # fflr 2.0.2